+++ packages/back-end/src/jobs/revalidateEventForwarderDataSourceQueries.ts
@@ -1,6 +1,6 @@
 import Agenda, { Job } from "agenda";
 import { getContextForAgendaJobByOrgId } from "back-end/src/services/organizations";
-import { revalidateManagedEventForwarderDataSourceQueries } from "back-end/src/services/eventForwarderWarehouseSync";
+import { revalidateManagedEventForwarderDataSourceQueries } from "back-end/src/services/eventForwarder/warehouseSync";
 import { logger } from "back-end/src/util/logger";
 
 const JOB_NAME = "revalidateEventForwarderDataSourceQueries";
