+++ packages/back-end/src/services/eventForwarder/bigquery.ts
@@ -2,7 +2,12 @@ import * as bq from "@google-cloud/bigquery";
 import {
   EVENT_FORWARDER_EXPERIMENT_VIEWED_TABLE,
   EVENT_FORWARDER_AVRO_PARTITION_FIELD,
+  isValidBigQueryTableName,
+  normalizeBigQueryTableNameForEventForwarder,
 } from "shared/util";
+import { BigQueryEventForwarderStoredConfig } from "shared/types/event-forwarder";
+import { EventForwarderConfigInterface } from "shared/validators";
+import { decryptEventForwarderConfigModel } from "back-end/src/services/eventForwarder/config";
 import { logger } from "back-end/src/util/logger";
 
 const FEATURE_USAGE_TABLE = "feature_usage";
@@ -56,6 +61,18 @@ type ServiceAccountKey = {
   client_email?: string;
   private_key?: string;
 };
+
+function validateBigQueryTableName(tableName: string): void {
+  if (!tableName) {
+    throw new Error("Missing BigQuery event forwarder table name");
+  }
+
+  if (!isValidBigQueryTableName(tableName)) {
+    throw new Error(
+      "Event forwarder table name must be a valid BigQuery table name (letters, numbers, underscores; Unicode letters allowed).",
+    );
+  }
+}
 
 function parseServiceAccountKey(raw: string): ServiceAccountKey {
   const trimmed = raw.trim();
@@ -116,6 +133,34 @@ async function ensureTable(
     { dataset: dataset.id, tableName },
     "Event forwarder BigQuery table created",
   );
+}
+
+/**
+ * Resolves the BigQuery table name for the Confluent sink. Existing tables are reused.
+ */
+export async function resolveBigQueryEventForwarderTableName(
+  eventForwarderConfig: EventForwarderConfigInterface,
+): Promise<string> {
+  const storedConfig =
+    decryptEventForwarderConfigModel<BigQueryEventForwarderStoredConfig>(
+      eventForwarderConfig,
+    );
+
+  const trimmed = storedConfig.tableName.trim();
+  if (!trimmed) {
+    throw new Error("Missing BigQuery event forwarder table name");
+  }
+
+  const baseTableName = normalizeBigQueryTableNameForEventForwarder(trimmed);
+
+  if (!storedConfig.dataset) {
+    throw new Error(
+      "Missing BigQuery dataset needed for connector provisioning",
+    );
+  }
+
+  validateBigQueryTableName(baseTableName);
+  return baseTableName;
 }
 
 export type EnsureEventForwarderBigQueryTablesParams = {
