+++ packages/back-end/test/services/eventForwarderConfig.draftUnchanged.test.ts
@@ -5,7 +5,7 @@ import {
   isEventForwarderDraftUnchanged,
   stripEventForwarderConfigMetadata,
   toEventForwarderConfigDraft,
-} from "back-end/src/services/eventForwarderConfig";
+} from "back-end/src/services/eventForwarder/config";
 
 const ENCRYPTION_KEY = "test-encryption-key-for-event-forwarder!!";
