+++ packages/front-end/components/Settings/EditDataSource/DataSourceInlineEditIdentifierTypes/EditIdentifierType.tsx
@@ -14,7 +14,8 @@ type EditIdentifierTypeProps = {
   userIdType: string;
   description?: string;
   attributes?: string[];
-  descriptionOnly?: boolean;
+  /** Event forwarder provisions hash-attribute identifier types; only description is editable. */
+  isEventForwarderManagedType?: boolean;
   onSave: (
     name: string,
     description: string,
@@ -28,7 +29,7 @@ export const EditIdentifierType: FC<EditIdentifierTypeProps> = ({
   userIdType,
   description,
   attributes,
-  descriptionOnly = false,
+  isEventForwarderManagedType = false,
   onSave,
   onCancel,
 }) => {
@@ -112,7 +113,7 @@ export const EditIdentifierType: FC<EditIdentifierTypeProps> = ({
           label="Identifier Type"
           {...form.register("idType")}
           pattern="^[a-z_]+$"
-          readOnly={mode === "edit" || descriptionOnly}
+          readOnly={mode === "edit" || isEventForwarderManagedType}
           required
           error={fieldError}
           helpText="Only lowercase letters and underscores allowed. For example, 'user_id' or 'device_cookie'."
@@ -125,7 +126,7 @@ export const EditIdentifierType: FC<EditIdentifierTypeProps> = ({
           maxRows={5}
           textarea
         />
-        {hashAttributes && !descriptionOnly && (
+        {hashAttributes && !isEventForwarderManagedType && (
           <MultiSelectField
             label="Hash Attributes"
             value={form.watch("attributes")}
